Refactored statistics logic into model
