--- .idea/libraries/lib.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="lib">
-    <CLASSES>
-      <root url="file://$PROJECT_DIR$/../DAT216/javafx-sdk-17.0.2/lib" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-    <jarDirectory url="file://$PROJECT_DIR$/../DAT216/javafx-sdk-17.0.2/lib" recursive="false" />
-  </library>
-</component>
