+++ TDA367-Journal.iml
@@ -7,7 +7,6 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="lib" level="project" />
-    <orderEntry type="library" name="JavaFX" level="project" />
+    <orderEntry type="library" exported="" name="javafx-sdk-17.0" level="application" />
   </component>
 </module>
