+++ src/MVC/Controller.java
@@ -90,13 +90,57 @@ public class Controller {
     @FXML private RadioButton passwordLockRadioButton;
     @FXML private RadioButton patternLockRadioButton;
     @FXML private RadioButton pinLockRadioButton;
-
-    @FXML private Slider moodSliderFour;
+/*
+Mood sliders
+1:
+2:
+3:
+4:
+* */
     @FXML private Slider moodSliderOne;
-    @FXML private Slider moodSliderThree;
     @FXML private Slider moodSliderTwo;
+    @FXML private Slider moodSliderThree;
+    @FXML private Slider moodSliderFour;
     @FXML private ToggleGroup Lock;
     @FXML private ToggleGroup Grade;
+
+
+
+    @FXML
+    private TabPane statisticsTabPane;
+    @FXML
+    private Tab statisticsGradeTab;
+    @FXML
+    private LineChart<LocalDatej, Integer> statisticsGradeTabChart;
+    @FXML
+    private Button statisticsGradeTabWeekBtn;
+    @FXML
+    private Button statisticsGradeTabMonthBtn;
+    @FXML
+    private Button statisticsGradeTabYearBtn;
+    @FXML
+    private Tab statisticsMoodTab;
+    @FXML
+    private LineChart<LocalDate, Integer> statisticsMoodTabChart;
+    @FXML
+    private Button statisticsMoodTabWeekBtn;
+    @FXML
+    private Button statisticsMoodTabMonthBtn;
+    @FXML
+    private Button statisticsMoodTabYearBtn;
+    @FXML
+    private Tab statisticsConditionTab;
+    @FXML
+    private PieChart statisticsConditionTabChart;
+    @FXML
+    private Tab statisticsTagTab;
+    @FXML
+    private PieChart statisticsTagTabChart;
+
+
+
+
+
     @FXML private TextField dailyPostTextField = new TextField();
     @FXML private DatePicker datePicker;
 
@@ -335,200 +379,26 @@ public class Controller {
     //Assume functionality exists in model
 
     //This method returns a hashmap with the correct dates and values that will be in the x and y axis of the chart respectively.
-    Pair<ArrayList<LocalDate>,ArrayList<HashMap<String,Integer>>>  intervalToDataMap(ETimeInterval ti){
-         HashMap<LocalDate,HashMap<String,Integer>> dates = new HashMap<>();
-         Pair<ArrayList<LocalDate>,ArrayList<HashMap<String,Integer>>> dataPair = new Pair<>(new ArrayList<>(),new ArrayList<>());
-        if(ti==null) return dataPair;
-        switch (ti){
-            case WEEK -> {
-                LocalDate currentDate = LocalDate.now();
-                for (int i = 0; i < 7; i++) {
-                    LocalDate d = currentDate.minusDays(i);
-                    DailyPost dailyPost = (DailyPost) model.getPosts().get(d);
-                    if (dailyPost == null){
-                        dataPair.getKey().add(d);
-                        dataPair.getValue().add(null);
-                        continue;
-                    }
-                    HashMap<String,Integer> tmpMap = new HashMap<>();
-                    dailyPost.getActiveMoods().forEach(e->{
-                        tmpMap.put(e.getMoodName(),e.getMoodRating());
-                    });
-                    dates.put(d,tmpMap);
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(tmpMap);
-                }
 
-                break;
-            }
-            case MONTH -> {
-                LocalDate currentDate = LocalDate.now();
-                for (int i = 0; i < currentDate.getMonth().length(currentDate.isLeapYear()); i++) {
-                    LocalDate d = currentDate.minusDays(i);
-                    DailyPost dailyPost = (DailyPost) model.getPosts().get(d);
-                    if (dailyPost == null){
-                        dataPair.getKey().add(d);
-                        dataPair.getValue().add(null);
-                        continue;
-                    }
-                    HashMap<String, Integer> tmpMap= new HashMap<>();
-                    model.getPosts().get(d).getActiveMoods().forEach(e->{
-                        tmpMap.put(e.getMoodName(),e.getMoodRating());
-                    });
-
-                    dates.put(d,tmpMap);
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(tmpMap);
-                }
-                break;
-            }
-            case YEAR -> {
-
-                for (int i = 0; i < 52; i++) {
-                    LocalDate d = LocalDate.now().minusWeeks(i);
-                    /*HashMap<EMood,IMood> weeklyMoodMap;*/
-                    HashMap<String, Integer> weeklyMoodMap = new HashMap<>() ;
-                    HashMap<String, Pair<Integer,Integer>> tmpMoodMap = new HashMap<>() ;
-                    IMood m = null;
-                    for (int j = 0; j < 7; j++) {
-                       LocalDate currentDay =  d.minusDays(d.getDayOfWeek().getValue()-1 - j);
-                        DailyPost dailyPost = (DailyPost) model.getPosts().get(currentDay);
-                        if (dailyPost == null){
-
-                            continue;
-                        }
-
-                        dailyPost.getActiveMoods().forEach(e -> {
-                            tmpMoodMap.putIfAbsent(e.getMoodName(),new Pair<>(0,0));
-                            Pair<Integer, Integer> p = tmpMoodMap.get(e.getMoodName());
-                            tmpMoodMap.put(e.getMoodName(),
-                                    new Pair<>(p.getKey()+1,p.getValue().intValue()+e.getMoodRating())) ;
-                        });
-
-                    }
-                    tmpMoodMap.forEach((k,v)-> weeklyMoodMap.put(k,v.getValue()/v.getKey()));
-                    dates.put(d, weeklyMoodMap);
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(weeklyMoodMap);
-                    System.out.println(weeklyMoodMap );
-                }
-                System.out.println(dates);
-                break;
-            }
-
-
-        }
-
-         System.out.println("dataDates: " + dataPair.getValue().size()+ " dataValues: " + dataPair.getKey().size());
-
-        return dataPair;
-    }
     private void populateMoodChart(){
         ETimeInterval t = ETimeInterval.YEAR;
-        Pair<ArrayList<LocalDate>,ArrayList<HashMap<String,Integer>>> data = intervalToDataMap(t);
+        Pair<ArrayList<LocalDate>,ArrayList<HashMap<String,Integer>>> data = model.intervalToDataMap(t);
 
     }
     private void populateConditionChart(){
-        Map<ECondition, Integer> conditionCountMap = new HashMap<>();
-        for (ECondition e:ECondition.values()) {
-            conditionCountMap.put(e,0);
-        }
-        model.getPosts().values().stream().map(e->{
-            List<ECondition> conditions = e.getConditions();
-            for (ECondition eCondition: conditions) {
-                int x = conditionCountMap.get(eCondition).intValue();
-                conditionCountMap.put(eCondition, x++);
-            }
-            return null;
-        });
+        Map<ECondition, Integer> conditionCountMap = model.getConditionData();
 
 
 
     }
     public void populateGradeChart(){
         ETimeInterval ti = ETimeInterval.MONTH;
-        Pair<ArrayList<LocalDate>,ArrayList<Integer>> pair =intervalToGradeData(ti);
+        Pair<ArrayList<LocalDate>,ArrayList<Integer>> pair =model.intervalToGradeData(ti);
+        
 
 
     }
-    Pair<ArrayList<LocalDate>,ArrayList<Integer>>  intervalToGradeData(ETimeInterval ti){
-        HashMap<LocalDate,HashMap<String,Integer>> dates = new HashMap<>();
-        Pair<ArrayList<LocalDate>,ArrayList<Integer>> dataPair = new Pair<>(new ArrayList<>(),new ArrayList<>());
-        if(ti==null) return dataPair;
-        switch (ti){
-            case WEEK -> {
-                LocalDate currentDate = LocalDate.now();
-                for (int i = 0; i < 7; i++) {
-                    LocalDate d = currentDate.minusDays(i);
-                    DailyPost dailyPost = (DailyPost) model.getPosts().get(d);
-                    if (dailyPost == null){
-                        dataPair.getKey().add(d);
-                        dataPair.getValue().add(null);
-                        continue;
-                    }
 
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(dailyPost.getGrade());
-                }
-
-                break;
-            }
-            case MONTH -> {
-                LocalDate currentDate = LocalDate.now();
-                for (int i = 0; i < currentDate.getMonth().length(currentDate.isLeapYear()); i++) {
-                    LocalDate d = currentDate.minusDays(i);
-                    DailyPost dailyPost = (DailyPost) model.getPosts().get(d);
-                    if (dailyPost == null){
-                        dataPair.getKey().add(d);
-                        dataPair.getValue().add(null);
-                        continue;
-                    }
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(dailyPost.getGrade());
-
-
-                }
-                break;
-            }
-            case YEAR -> {
-
-                for (int i = 0; i < 52; i++) {
-                    LocalDate d = LocalDate.now().minusWeeks(i);
-                    /*HashMap<EMood,IMood> weeklyMoodMap;*/
-                    Integer weeklyRating = null;
-                    Pair<Integer,Integer> tmpWeeklyRating = new Pair<>(0,0);
-                    IMood m = null;
-                    for (int j = 0; j < 7; j++) {
-
-                        LocalDate currentDay =  d.minusDays(d.getDayOfWeek().getValue()-1 - j);
-                        DailyPost dailyPost = (DailyPost) model.getPosts().get(currentDay);
-                        if (dailyPost == null || dailyPost.getGrade() == 0){
-                            continue;
-                        }
-
-                        tmpWeeklyRating = new Pair<>(tmpWeeklyRating.getKey()+1,tmpWeeklyRating.getValue()+ dailyPost.getGrade());
-
-
-                    }
-                    if (tmpWeeklyRating.getKey()==0){
-                        dataPair.getKey().add(d);
-                        dataPair.getValue().add(0);
-                    }
-                    weeklyRating = tmpWeeklyRating.getValue()/tmpWeeklyRating.getKey();
-
-                    dataPair.getKey().add(d);
-                    dataPair.getValue().add(weeklyRating);
-                }
-                break;
-            }
-
-
-        }
-
-        System.out.println("dataDates: " + dataPair.getValue()+ " dataValues: " + dataPair.getKey());
-
-        return dataPair;
-    }
    /* public void populateDayRatingChart(){
         //this code below
         *//*Map<Integer, Long> chartData = model.getPosts().values().stream()
@@ -552,15 +422,7 @@ public class Controller {
 
     }*/
     private void populateTagsChart(){
-        ArrayList<ITag> tags = new ArrayList<>(model.getAllTags());
-        /*HashMap<LocalDate, ArrayList<IMood>> tagsMap = model.getTagsMap();*/
-        Map<Object, Long> chartData = model.getPosts().values().stream()
-                .collect(Collectors.groupingBy(p -> p.getTags(),
-                        Collectors.counting()));
-        for (int x = 0; x < tags.size(); x++){
-            chartData.putIfAbsent(tags.get(x),0L);
-        }
-        System.out.println(tags);
+        model.getTagData();
     }
     //-----------------------Statistics logic end-----------------------
 }
